오브젝트 생성, 삭제, 이동 요청 구현

+++ Swarm_Forest/Assets/Scripts/Skill/Skill.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class Skill: MonoBehaviour{
         
     public int ID;    
-    public float BaseDamage;    
+    public int BaseDamage;    
     public float PercentDamage;    
     public float LifeTime;    
     public float Cooldown;
